images
comment on the images: please make new ones, the current ones are ok as placeholders, but make the whole thing look unprofessional. They should be gone when this is launched. Sascha will be a good person to talk to here.
comment on the text: would leave out the doubling part and be less specific on the number of employees